Move name and description into commons folder
