+++ src/main/java/seedu/address/model/commons/Description.java
@@ -1,10 +1,10 @@
-package seedu.address.model.attraction;
+package seedu.address.model.commons;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
 
 /**
- * Represents a Attraction's description in TrackPad.
+ * Represents an item's description in TrackPad.
  * Guarantees: immutable; is valid as declared in {@link #isValidDescription(String)}
  */
 public class Description {
